update mysql root

+++ bamazon.sql
@@ -3,7 +3,7 @@ CREATE DATABASE bamazon;
 USE bamazon;
 
 CREATE TABLE products (
-	item_id INTEGER(11) AUTO_INCREMENT NOT NULL,
+	item_id INTEGER(20) AUTO_INCREMENT NOT NULL,
 	product_name VARCHAR(30) NOT NULL,
 	department_name VARCHAR(20) NOT NULL,
 	price DECIMAL(10,2) NOT NULL,
